anaemia_detection/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
Pillow>=8.3,<9.0
tensorflow>=2.6,<3.0
numpy>=1.21,<2.0
